Adding tolerance to nonexistent files
As it is, the overlay system will refuse to build a commit if it doesn't
present a file in each category `*.$OVERLAY` and `bin/*.$OVERLAY`. This
causes unfortunate effects when some of the files are unnecessary for
the commit.

This adds support for the nonexistent-files case; the script will
continue to build normally even if some of the files don't actually exist.